+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [GH#166](https://github.com/jolicode/automapper/pull/166) Remove cache for property info, use specific services instead
 
 ## [9.1.1] - 2024-06-19
 ### Fixed