+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [9.1.2] - 2024-09-03
 ### Fixed
 - [GH#174](https://github.com/jolicode/automapper/pull/174) Fix race condition when writing generated mappers
 - [GH#167](https://github.com/jolicode/automapper/pull/167) Fix property metadata attribute name in docs
@@ -329,7 +331,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * [AutoMapper] [GH#179](https://github.com/janephp/janephp/pull/179) Fixing incompatible changes in Symfony 5.0
 
-[Unreleased]: https://github.com/jolicode/automapper/compare/9.1.1...HEAD
+[Unreleased]: https://github.com/jolicode/automapper/compare/9.1.2...HEAD
+[9.1.2]: https://github.com/janephp/janephp/compare/9.1.1...9.1.2
 [9.1.1]: https://github.com/janephp/janephp/compare/9.1.0...9.1.1
 [9.1.0]: https://github.com/janephp/janephp/compare/9.0.2...9.1.0
 [9.0.2]: https://github.com/janephp/janephp/compare/9.0.1...9.0.2